more refactoring...
