+++ src/main/java/com/smartbear/readyapi/client/teststeps/restrequest/RestRequestStepBuilder.java
@@ -1,36 +1,98 @@
 package com.smartbear.readyapi.client.teststeps.restrequest;
 
+import com.smartbear.readyapi.client.model.Parameter;
+import com.smartbear.readyapi.client.model.RestTestRequestStep;
+import com.smartbear.readyapi.client.teststeps.TestStepTypes;
+import com.smartbear.readyapi.client.teststeps.TestSteps;
 import com.smartbear.readyapi.client.teststeps.request.HttpRequestStepBuilder;
 
-public interface RestRequestStepBuilder<RestRequestBuilderType extends RestRequestStepBuilder> extends HttpRequestStepBuilder<RestRequestBuilderType> {
+import static com.smartbear.readyapi.client.Validator.validateNotEmpty;
+import static com.smartbear.readyapi.client.assertions.Assertions.jsonPathContent;
+import static com.smartbear.readyapi.client.assertions.Assertions.jsonPathCount;
 
-    RestRequestBuilderType addQueryParameter(String parameterName, String value);
+public class RestRequestStepBuilder<RestRequestBuilderType extends RestRequestStepBuilder> extends HttpRequestStepBuilder<RestRequestBuilderType, RestTestRequestStep> {
 
-    RestRequestBuilderType addPathParameter(String parameterName, String value);
+    public enum ParameterType {
+        MATRIX, HEADER, QUERY, PATH
+    }
 
-    RestRequestBuilderType addMatrixParameter(String parameterName, String value);
+    public RestRequestStepBuilder(String uri, TestSteps.HttpMethod method) {
+        super(new RestTestRequestStep(), TestStepTypes.REST_REQUEST.getName());
+        testStep.setURI(uri);
+        testStep.setMethod(method.toString());
+    }
 
-    RestRequestBuilderType addHeaderParameter(String parameterName, String value);
+    public RestRequestBuilderType addQueryParameter(String parameterName, String value) {
+        return addParameter(parameterName, value, ParameterType.QUERY);
+    }
 
-    RestRequestBuilderType postQueryString();
+    public RestRequestBuilderType addPathParameter(String parameterName, String value) {
+        return addParameter(parameterName, value, ParameterType.PATH);
+    }
 
-    /**
-     * Assertion shortcuts
-     */
+    public RestRequestBuilderType addMatrixParameter(String parameterName, String value) {
+        return addParameter(parameterName, value, ParameterType.MATRIX);
+    }
 
-    RestRequestBuilderType assertJsonContent(String jsonPath, String expectedContent);
+    public RestRequestBuilderType addHeaderParameter(String parameterName, String value) {
+        return addParameter(parameterName, value, ParameterType.HEADER);
+    }
 
-    RestRequestBuilderType assertJsonCount(String jsonPath, int expectedCount);
+    protected RestRequestBuilderType addParameter(String parameterName, String value, RestRequestStepBuilder.ParameterType type) {
+        Parameter parameter = new Parameter();
+        parameter.setName(parameterName);
+        parameter.setValue(value);
+        parameter.setType(type.name());
 
-    RestRequestBuilderType assertValidStatusCodes(Integer... statusCodes);
+        getParameters().add(parameter);
+        return (RestRequestBuilderType) this;
+    }
 
-    RestRequestBuilderType assertInvalidStatusCodes(Integer... statusCodes);
+    public RestRequestBuilderType postQueryString() {
+        testStep.setPostQueryString(true);
+        return (RestRequestBuilderType) this;
+    }
 
-    RestRequestBuilderType get(String uri);
+    public RestRequestBuilderType assertJsonContent(String jsonPath, String expectedContent) {
+        return addAssertion(jsonPathContent(jsonPath, expectedContent).allowWildcards());
+    }
 
-    RestRequestBuilderType post(String uri);
+    public RestRequestBuilderType assertJsonCount(String jsonPath, int expectedCount) {
+        return addAssertion(jsonPathCount(jsonPath, expectedCount).allowWildcards());
+    }
 
-    RestRequestBuilderType put(String uri);
+    public RestRequestBuilderType get(String uri) {
+        testStep.setMethod("GET");
+        testStep.setURI(uri);
 
-    RestRequestBuilderType delete(String uri);
+        return (RestRequestBuilderType) this;
+    }
+
+    public RestRequestBuilderType post(String uri) {
+        testStep.setMethod("POST");
+        testStep.setURI(uri);
+
+        return (RestRequestBuilderType) this;
+    }
+
+    public RestRequestBuilderType put(String uri) {
+        testStep.setMethod("PUT");
+        testStep.setURI(uri);
+
+        return (RestRequestBuilderType) this;
+    }
+
+    public RestRequestBuilderType delete(String uri) {
+        testStep.setMethod("DELETE");
+        testStep.setURI(uri);
+
+        return (RestRequestBuilderType) this;
+    }
+
+    public RestTestRequestStep build() {
+        super.build();
+        validateNotEmpty(testStep.getURI(), "No URI set, it's a mandatory parameter for REST Request");
+        validateNotEmpty(testStep.getMethod(), "No HTTP method set, it's a mandatory parameter for REST Request");
+        return testStep;
+    }
 }
