+++ src/main/java/com/smartbear/readyapi/client/teststeps/restrequest/RestRequestStepWithBodyBuilder.java
@@ -1,9 +1,26 @@
 package com.smartbear.readyapi.client.teststeps.restrequest;
 
-public interface RestRequestStepWithBodyBuilder extends RestRequestStepBuilder<RestRequestStepWithBodyBuilder> {
-    RestRequestStepWithBodyBuilder withRequestBody(String requestBody);
+import com.smartbear.readyapi.client.teststeps.TestSteps;
 
-    RestRequestStepWithBodyBuilder withMediaType(String mediaType);
+public class RestRequestStepWithBodyBuilder extends RestRequestStepBuilder<RestRequestStepWithBodyBuilder> {
 
-    RestRequestStepWithBodyBuilder withEncoding(String encoding);
+    public RestRequestStepWithBodyBuilder(String uri, TestSteps.HttpMethod post) {
+        super(uri, post);
+    }
+
+    public RestRequestStepWithBodyBuilder withRequestBody(String requestBody) {
+        testStep.setRequestBody(requestBody);
+        return this;
+    }
+
+    public RestRequestStepWithBodyBuilder withMediaType(String mediaType) {
+        testStep.setMediaType(mediaType);
+        return this;
+    }
+
+    public RestRequestStepWithBodyBuilder withEncoding(String encoding) {
+        testStep.setEncoding(encoding);
+        return this;
+    }
+
 }
