--- src/main/java/com/smartbear/readyapi/client/teststeps/restrequest/RestRequestWithBodyBuilder.java
@@ -1,26 +1,0 @@
-package com.smartbear.readyapi.client.teststeps.restrequest;
-
-import com.smartbear.readyapi.client.teststeps.TestSteps;
-
-public class RestRequestWithBodyBuilder extends BaseRestRequestStepBuilder<RestRequestStepWithBodyBuilder> implements RestRequestStepWithBodyBuilder {
-
-    public RestRequestWithBodyBuilder(String uri, TestSteps.HttpMethod post) {
-        super(uri, post);
-    }
-
-    public RestRequestStepWithBodyBuilder withRequestBody(String requestBody) {
-        testStep.setRequestBody(requestBody);
-        return this;
-    }
-
-    public RestRequestStepWithBodyBuilder withMediaType(String mediaType) {
-        testStep.setMediaType(mediaType);
-        return this;
-    }
-
-    public RestRequestStepWithBodyBuilder withEncoding(String encoding) {
-        testStep.setEncoding(encoding);
-        return this;
-    }
-
-}
