+++ src/test/java/com/smartbear/readyapi/client/RestRequestStepRecipeTest.java
@@ -21,9 +21,9 @@ import static com.smartbear.readyapi.client.teststeps.TestSteps.getRequest;
 import static com.smartbear.readyapi.client.teststeps.TestSteps.postRequest;
 import static com.smartbear.readyapi.client.teststeps.TestSteps.putRequest;
 import static com.smartbear.readyapi.client.teststeps.TestSteps.restRequest;
-import static com.smartbear.readyapi.client.teststeps.restrequest.BaseRestRequestStepBuilder.ParameterType.HEADER;
-import static com.smartbear.readyapi.client.teststeps.restrequest.BaseRestRequestStepBuilder.ParameterType.MATRIX;
-import static com.smartbear.readyapi.client.teststeps.restrequest.BaseRestRequestStepBuilder.ParameterType.QUERY;
+import static com.smartbear.readyapi.client.teststeps.restrequest.RestRequestStepBuilder.ParameterType.HEADER;
+import static com.smartbear.readyapi.client.teststeps.restrequest.RestRequestStepBuilder.ParameterType.MATRIX;
+import static com.smartbear.readyapi.client.teststeps.restrequest.RestRequestStepBuilder.ParameterType.QUERY;
 import static java.util.Collections.singletonList;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
